Move _from_metadata to store
